Python-Projects-TanCodes/RepoLoader/requirements.txt
pandas==1.2.2
streamlit==1.2.0 
st-annotated-text==2.0.0
Pillow==8.0.1
base58==2.1.0
requests==2.25.1
beautifulsoup4==4.9.3
